# src/lib/openapi.yaml
openapi: 3.0.0
info:
  title: PetStayTion API
  version: 1.0.0
  description: API for pet care service management
servers:
  - url: /
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/pets:
    get:
      summary: Get all pets for the authenticated user
      description: |
        Returns all pets for the authenticated user associated with any of their clients.
        Requires a valid Firebase authentication token.
      tags: [Pets]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Pets retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: "Max"
                        age:
                          type: number
                          example: 5
                        description:
                          type: string
                          nullable: true
                          example: "Friendly dog"
                        keyImage:
                          type: string
                          nullable: true
                          example: "https://example.com/pet.jpg"
                        species:
                          type: string
                          example: "Dog"
        "401":
          description: Unauthorized (invalid or missing token)
        "404":
          description: User not found
        "500":
          description: Internal server error
    post:
      summary: Create a new pet
      description: |
        Creates a new pet for the authenticated user.
        Species is created if it doesn't exist.
        Requires a valid Firebase authentication token.
        At least one image must be provided.
      tags: [Pets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - age
                - images
                - speciesName
              properties:
                name:
                  type: string
                  example: "Max"
                age:
                  type: number
                  minimum: 0
                  maximum: 999
                  example: 5
                description:
                  type: string
                  nullable: true
                  example: "Friendly dog"
                images:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: binary
                speciesName:
                  type: string
                  example: "Dog"
      responses:
        "200":
          description: Pet created successfully
        "400":
          description: Missing required fields, no images provided, or invalid age
        "401":
          description: Unauthorized (invalid or missing token)
        "404":
          description: User or client not found
        "409":
          description: Conflict (pet with this name already exists for the user)
        "500":
          description: Internal server error
...
  /api/pets/{id}:
    get:
      summary: Get a pet by ID
      description: Returns detailed information about a specific pet including images and species.
      tags: [Pets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The pet ID
      responses:
        "200":
          description: Pet details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pet:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "Max"
                      age:
                        type: number
                        example: 5
                      description:
                        type: string
                        nullable: true
                        example: "Friendly dog"
                      species:
                        type: string
                        example: "Dog"
                      images:
                        type: array
                        items:
                          type: object
                          properties:
                            imageUrl:
                              type: string
                              example: "https://example.com/pet.jpg"
                            order:
                              type: integer
                              nullable: true
                              example: 1
        "400":
          description: Invalid pet ID
        "404":
          description: Pet not found
        "500":
          description: Internal server error
    put:
      summary: Update a pet
      description: |
        Updates an existing pet's information. Only the pet owner can update the pet.
        Requires a valid Firebase authentication token.
        Species is created if it doesn't exist.
        At least one image must be provided (new or kept).
      tags: [Pets]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The pet ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - age
                - speciesName
              properties:
                name:
                  type: string
                  example: "Max"
                age:
                  type: number
                  minimum: 0
                  maximum: 999
                  example: 5
                description:
                  type: string
                  nullable: true
                  example: "Friendly dog"
                speciesName:
                  type: string
                  example: "Dog"
                keepImageUrls:
                  type: string
                  description: JSON array of URLs to keep
                  example: '["https://example.com/old1.jpg"]'
                newImages:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Pet updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Missing required fields, no images provided, invalid pet ID, or invalid age
        "401":
          description: Unauthorized (invalid or missing token)
        "403":
          description: Forbidden (user is not the pet owner or not a client)
        "404":
          description: Pet not found or user not found
        "409":
          description: Conflict (pet with this name already exists for the user)
        "500":
          description: Internal server error
...